Big bang

+++ lib/eternity/session.rb
@@ -1,17 +1,19 @@
 module Eternity
   class Session
 
-    attr_reader :name, :index, :namespace
+    attr_reader :name, :index, :key, :branches
     
-    def initialize(name)
-      @name = name.to_s
-      @namespace = Eternity.namespace[:session][name]
+    def initialize(name=nil)
+      @name = name.to_s || Restruct.generate_key
+      @key = Eternity.keyspace[:session][@name]
       @index = Index.new self
       @delta = Delta.new self
+      @current = Restruct::Hash.new key: key[:current]
+      @branches = Restruct::Hash.new key: key[:branches]
     end
 
     def current_commit_id
-      Eternity.redis.call 'GET', namespace[:current_commit]
+      current[:commit]
     end
 
     def current_commit
@@ -19,27 +21,23 @@ module Eternity
     end
 
     def current_commit?
-      Eternity.redis.call('EXISTS', namespace[:current_commit]) == 1
+      current.key? :commit
     end
 
     def current_branch
-      Eternity.redis.call('GET', namespace[:current_branch]) || 'master'
-    end
-
-    def branches
-      Hash[Eternity.redis.call('HGETALL', namespace[:branches]).each_slice(2).to_a]
+      current[:branch] || 'master'
     end
 
     def entries
-      index.entries
+      index.to_h
     end
 
-    def delta
-      @delta.to_h
+    def changes
+      delta.to_h
     end
 
-    def delta?
-      !@delta.empty?
+    def changes?
+      !delta.empty?
     end
 
     def [](section)
@@ -47,61 +45,46 @@ module Eternity
     end
 
     def commit(options)
-      raise 'Nothing to commit' if @delta.empty?
-
-      params = {
-        parents: [current_commit_id].compact,
-        index: Blob.write(:index, index.dump),
-        delta: Blob.write(:delta, @delta.to_h)
-      }
-      
-      commit_id = Commit.create options.merge(params)
-
-      Eternity.redis.call 'SET', namespace[:current_commit], commit_id
-      Eternity.redis.call 'HSET', namespace[:branches], current_branch, commit_id
-
-      @delta.destroy
-
-      commit_id
+      commit! message: options.fetch(:message), 
+              author: options.fetch(:author)
     end
 
     def revert
-      @delta.destroy
+      delta.destroy
       index.revert
     end
 
     def branch(name)
       raise 'Cant branch without commit' unless current_commit?
-      raise 'Cant branch with uncommitted changes' if delta?
+      raise 'Cant branch with uncommitted changes' if changes?
 
-      Eternity.redis.call 'HSET', namespace[:branches], name, current_commit_id
+      branches[name] = current_commit_id
     end
 
     def checkout(branch)
-      raise 'Cant checkout with uncommitted changes' if delta?
+      raise 'Cant checkout with uncommitted changes' if changes?
 
       commit_id =
-        if branches.key?(branch.to_s)
-          branches[branch.to_s]
+        if branches.key? branch
+          branches[branch]
         elsif Branch.exists?(branch)
           Branch.new(branch).commit_id
         else
           raise "Invalid branch #{branch}"
         end
 
-      Eternity.redis.call 'SET', namespace[:current_commit], commit_id
-      Eternity.redis.call 'SET', namespace[:current_branch], branch
-      Eternity.redis.call 'HSET', namespace[:branches], branch, commit_id
+      current[:commit] = commit_id
+      current[:branch] = branch
+      branches[branch] = commit_id
       
-      @delta.destroy
-      index.destroy
+      delta.destroy
       index.restore Commit.new(commit_id).index_dump
     end
 
     def push
       raise 'Cant push without commit' unless current_commit?
       raise 'Push rejected. Non fast forward' unless current_commit.fast_forward?(Branch.new(current_branch).commit_id)
-
+      
       push!
     end
 
@@ -109,13 +92,56 @@ module Eternity
       Branch.new(current_branch).commit_id = current_commit_id
     end
 
-    def destroy
-      Eternity.redis.call 'DEL', namespace[:current_commit]
-      Eternity.redis.call 'DEL', namespace[:current_branch]
-      Eternity.redis.call 'DEL', namespace[:branches]
+    def pull
+      raise 'Cant pull with uncommitted changes' if changes?
+      raise "Branch not found: #{current_branch}" unless Branch.exists? current_branch
+      
+      branch = Branch.new current_branch
 
-      @delta.destroy
+      if branch.commit.fast_forward? current_commit_id
+        branches[current_branch] = branch.commit_id
+        checkout current_branch
+      else
+        patch = Patch.new current_commit, branch.commit
+        patch.apply_to index
+
+        commit! author: 'System',
+                message: "Merge #{branch.commit_id} into #{current_commit_id}",
+                parents: patch.commit_ids,
+                delta: Blob.write(:delta, {}),
+                base: patch.base.id,
+                base_delta: Blob.write(:delta, patch.base_delta)
+      end
+    end
+
+    def destroy
+      current.destroy
+      branches.destroy
+      delta.destroy
       index.destroy
+    end
+
+    def restore(commit)
+      delta.destroy
+      index.restore commit.index_dump
+    end
+
+    private
+
+    attr_reader :delta, :current
+
+    def commit!(options)
+      raise 'Nothing to commit' unless changes?
+
+      options[:index] = Blob.write :index, index.dump
+      options[:delta] ||= Blob.write :delta, delta.to_h
+      options[:parents] ||= [current_commit_id].compact
+      
+      Commit.create(options).tap do |commit_id|
+        current[:commit] = commit_id
+        branches[current_branch] = commit_id
+        delta.destroy
+      end
     end
 
   end
